Add ui.js, modularize code
